fixed casting issue, created new handin

+++ src/rasterizer/framebuffer.cpp
@@ -30,7 +30,7 @@ HDR_Image Framebuffer::resolve_colors() const {
 
 	for (uint32_t y = 0; y < height; ++y) {
 		for (uint32_t x = 0; x < width; ++x) {
-			for(int s = 0; s < sample_pattern.centers_and_weights.size(); s++){
+			for(int s = 0; s < (int) sample_pattern.centers_and_weights.size(); s++){
 				image.at(x,y) += color_at(x,y,s) * sample_pattern.centers_and_weights[s][2];
 			}
